refactor: Config & enum
