+++ api/src/main/kotlin/com/gotchai/api/global/config/SecurityConfig.kt
@@ -1,13 +1,6 @@
 package com.gotchai.api.global.config
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateSerializer
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalTimeSerializer
-import com.fasterxml.jackson.datatype.jsr310.ser.YearMonthSerializer
-import com.fasterxml.jackson.datatype.jsr310.ser.ZonedDateTimeSerializer
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -21,44 +14,11 @@ import org.springframework.security.crypto.factory.PasswordEncoderFactories
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.security.oauth2.jwt.Jwt
 import org.springframework.security.web.SecurityFilterChain
-import java.time.LocalDate
-import java.time.LocalDateTime
-import java.time.LocalTime
-import java.time.YearMonth
-import java.time.ZonedDateTime
-import java.time.format.DateTimeFormatter
 
 @Configuration
 @EnableWebSecurity
 @EnableConfigurationProperties(RsaKeyProperties::class)
 class SecurityConfig {
-    @Bean
-    fun objectMapper(): ObjectMapper =
-        jacksonObjectMapper().registerModules(
-            JavaTimeModule().apply {
-                addSerializer(
-                    LocalDate::class.java,
-                    LocalDateSerializer(DateTimeFormatter.ISO_DATE),
-                )
-                addSerializer(
-                    LocalDateTime::class.java,
-                    LocalDateTimeSerializer(DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS")),
-                )
-                addSerializer(
-                    ZonedDateTime::class.java,
-                    ZonedDateTimeSerializer(DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSX")),
-                )
-                addSerializer(
-                    LocalTime::class.java,
-                    LocalTimeSerializer(DateTimeFormatter.ofPattern("HH:mm:ss.SSS")),
-                )
-                addSerializer(
-                    YearMonth::class.java,
-                    YearMonthSerializer(DateTimeFormatter.ofPattern("yyyy-MM")),
-                )
-            },
-        )
-
     @Bean
     fun passwordEncoder(): PasswordEncoder = PasswordEncoderFactories.createDelegatingPasswordEncoder()
 
@@ -69,6 +29,7 @@ class SecurityConfig {
     fun securityFilterChain(
         http: HttpSecurity,
         jwtConverter: Converter<Jwt, out AbstractAuthenticationToken>,
+        objectMapper: ObjectMapper,
     ): SecurityFilterChain {
         http.oauth2ResourceServer {
             it.jwt { jwtConfigurer ->
@@ -82,7 +43,7 @@ class SecurityConfig {
             .csrf { it.disable() }
             .formLogin { it.disable() }
             .sessionManagement { it.sessionCreationPolicy(SessionCreationPolicy.STATELESS) }
-            .exceptionHandling { it.authenticationEntryPoint(CustomAuthenticationEntryPoint(objectMapper())) }
+            .exceptionHandling { it.authenticationEntryPoint(CustomAuthenticationEntryPoint(objectMapper)) }
 
         http.authorizeHttpRequests { authorize ->
             authorize
