+++ common/src/main/kotlin/com/gotchai/common/enum/auth/AuthenticationEntityStatus.kt
@@ -1,4 +1,4 @@
-package com.gotchai.storage.rdb.auth
+package com.gotchai.common.enum.auth
 
 enum class AuthenticationEntityStatus {
     ACTIVE,
