+++ common/src/main/kotlin/com/gotchai/common/enum/auth/AuthorityType.kt
@@ -1,4 +1,4 @@
-package com.gotchai.domain.auth
+package com.gotchai.common.enum.auth
 
 enum class AuthorityType {
     USER,
