+++ common/src/main/kotlin/com/gotchai/common/enum/auth/TokenStatus.kt
@@ -1,4 +1,4 @@
-package com.gotchai.domain.auth.token
+package com.gotchai.common.enum.auth
 
 enum class TokenStatus {
     ACTIVE,
