+++ domain/src/main/kotlin/com/gotchai/domain/auth/AuthenticationHistory.kt
@@ -1,15 +1,13 @@
 package com.gotchai.domain.auth
 
-import com.gotchai.domain.auth.token.NewToken
-import com.gotchai.domain.auth.token.Token
-import com.gotchai.domain.auth.token.TokenStatus
+import com.gotchai.common.enum.auth.TokenStatus
 import java.time.LocalDateTime
 
 class AuthenticationHistory {
     data class Creation(
         val userId: Long,
         val deviceId: String?,
-        val newToken: NewToken,
+        val token: Token,
         val status: TokenStatus,
     )
 
@@ -25,6 +23,6 @@ class AuthenticationHistory {
         val userId: Long,
         val deviceId: String?,
         val refreshToken: String,
-        val newToken: NewToken,
+        val token: Token,
     )
 }
