+++ domain/src/main/kotlin/com/gotchai/domain/auth/GrantedAuthority.kt
@@ -1,5 +1,7 @@
 package com.gotchai.domain.auth
 
+import com.gotchai.common.enum.auth.AuthorityType
+
 data class GrantedAuthority(
     val authorityType: AuthorityType,
 )
