--- domain/src/main/kotlin/com/gotchai/domain/auth/token/NewToken.kt
@@ -1,5 +1,0 @@
-package com.gotchai.domain.auth.token
-
-data class NewToken(
-    val token: Token,
-)
