--- domain/src/main/kotlin/com/gotchai/domain/auth/token/RefreshToken.kt
@@ -1,5 +1,0 @@
-package com.gotchai.domain.auth.token
-
-data class RefreshToken(
-    val token: String,
-)
