+++ storage/rdb/src/main/kotlin/com/gotchai/storage/rdb/user/repository/UserJpaRepository.kt
@@ -1,5 +1,6 @@
-package com.gotchai.storage.rdb.user
+package com.gotchai.storage.rdb.user.repository
 
+import com.gotchai.storage.rdb.user.entity.UserEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface UserJpaRepository : JpaRepository<UserEntity, Long>
